Automate package collection and fix bumping
This uses the helper scripts in {deb,rpm}/develop to gather all packages
that need to be bumped. This is then inserted as a matrix of packages to
bump.

The bumping jobs are now reduced to setup (install dependencies + git
checkout) and calling a bumping helper with the name and version.

It also fixes the rpmdev-packager to be a minimal script in itself. An
alias can't be used inside another script.

It also renames the workflow to bump_packages.yml since it now has no
notion of gems. It just understands packages and versions. All specifics
are inside the {deb,rpm}/develop branches inside helpers. This opens the
path to bumping other things, such as Python packages or Ansible
collections.

The only downside that remains is that every run generates a new commit
which is then force pushed. This causes CI to run for every branch which
can become quite expensive if it runs daily.